debugged

+++ src/main/java/com/co2fp/backend/backend/DTO/transport/TransportRegisterDTO.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class TransportRegisterDTO {
     private String transport_name;
-    private Integer fuel_factor;
-    private Integer emission_factor;
+    private Float fuel_factor;
+    private Float emission_factor;
 }
